Refactor application service
