+++ GitBlamePR/ApplicationService.swift
@@ -8,11 +8,6 @@
 
 import Foundation
 
-enum ApplicationServiceError: Error {
-    case unknown
-    case standardError(String)
-}
-
 class ApplicationService: ObservableObject {
     var fullPath: String = "" {
         didSet {
@@ -41,13 +36,13 @@ class ApplicationService: ObservableObject {
         URL(fileURLWithPath: trimedFullPath).deletingLastPathComponent()
     }
 
-    func execute() {
+    private func execute() {
         var remoteOut = ""
         var blamePROut = ""
         do {
             remoteOut = try executeGitRemote()
             blamePROut = try executeGitBlamePR()
-        } catch ApplicationServiceError.standardError(let description) {
+        } catch ProcessError.standardError(let description) {
             viewModel = GitBlamePRViewModel()
             viewModel.error = description
             return
@@ -63,7 +58,7 @@ class ApplicationService: ObservableObject {
     }
 
     private func executeGitRemote() throws -> String {
-        return try runProcess(
+        return try Process.run(
             executableURL: URL(fileURLWithPath: "/usr/bin/git"),
             arguments: ["remote", "-v"],
             currentDirectoryURL: fullPathDirectoryURL
@@ -71,74 +66,10 @@ class ApplicationService: ObservableObject {
     }
 
     private func executeGitBlamePR() throws -> String {
-        return try runProcess(
+        return try Process.run(
             executableURL: URL(fileURLWithPath: "/usr/bin/perl"),
             arguments: [Bundle.main.resourcePath! + "/git-blame-pr.pl", trimedFullPath],
             currentDirectoryURL: fullPathDirectoryURL
         )
-    }
-
-    private func runProcess(executableURL: URL, arguments: [String], currentDirectoryURL: URL?) throws -> String {
-        let process = Process()
-        let stdOutput = Pipe()
-        let stdError = Pipe()
-        process.executableURL = executableURL
-        process.arguments = arguments
-        process.currentDirectoryURL = currentDirectoryURL
-        process.standardOutput = stdOutput
-        process.standardError = stdError
-        try process.run()
-        guard let errOut = String(data: stdError.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) else {
-            throw ApplicationServiceError.unknown
-        }
-        guard errOut.isEmpty else {
-            throw ApplicationServiceError.standardError(errOut)
-        }
-        guard let stdOut = String(data: stdOutput.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) else {
-            throw ApplicationServiceError.unknown
-        }
-        return stdOut
-    }
-
-}
-
-extension GitBlamePRViewModel {
-    init() {
-        self.lines = []
-        self.recent = RecentViewModel(fullPaths: [])
-        self.error = ""
-    }
-
-    init?(gitRemoteStandardOutput: String, gitBlamePRStandardOutput: String) {
-        guard let repo = gitRemoteStandardOutput
-            .components(separatedBy: .newlines)[0]
-            .components(separatedBy: "github.com:")
-            .last?
-            .components(separatedBy: ".git")
-            .first else { return nil }
-        let repoURL = "https://github.com/" + repo
-
-        let strLines = gitBlamePRStandardOutput.components(separatedBy: .newlines)
-        self.lines = strLines.map { (line) -> (message: String, url: URL, code: String, id: UUID) in
-            let linePrefix = line.prefix(9)
-            let code = line.suffix(line.count - linePrefix.count)
-            let trimdlinePrefix = linePrefix.trimmingCharacters(in: .whitespacesAndNewlines)
-            let message = trimdlinePrefix.replacingOccurrences(of: "^", with: "")
-            return (
-                message: message,
-                url: Self.url(message: message, repoURL: repoURL),
-                code: String(code),
-                id: UUID()
-            )
-        }
-        self.recent = RecentViewModel(fullPaths: [])
-    }
-
-    private static func url(message: String, repoURL: String) -> URL {
-        if message.contains("PR") {
-            let prNumber = message.components(separatedBy: "#").last!
-            return URL(string: repoURL + "/pull/" + prNumber)!
-        }
-        return URL(string: repoURL + "/commit/" + message)!
     }
 }
